backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
youtube-transcript-api
yt-dlp
langchain
langchain-community
groq
aiosqlite
langdetect